Code review suggestions — changing colors and moving file
